requirements.txt
-i https://pypi.org/simple
audioread==2.1.8
cffi==1.13.2
decorator==4.4.1
joblib==0.14.1
librosa==0.7.1
llvmlite==0.30.0
mido==1.2.9
numba==0.46.0
numpy==1.17.4 ; python_version >= '3.0'
pillow==6.2.1
pycparser==2.19
pyglm==1.1.2
pyo==1.0.1
pyopengl==3.1.4
resampy==0.2.2
scikit-learn==0.22
scipy==1.3.3
six==1.13.0
soundfile==0.10.3.post1
wxpython==4.0.7.post2
